hopefully the vercel will go through now. some type corrections for the sky
