--- src/Realm/properties/RealmScene/Matrix/components/matrix2.ts
@@ -1,168 +1,0 @@
-import { DoubleSide, ShaderMaterial, Uniform } from "three";
-import { useMemo } from "react";
-import { useLimiter } from "spacesvr";
-import { useFrame, useLoader } from "@react-three/fiber";
-import * as THREE from "three";
-import { useRealm } from "../../../../components/RealmState";
-
-export const useMatrixMat2 = (color: string): ShaderMaterial => {
-  const { effects = { name: "Fog", color: "Black" } } = useRealm();
-  // const colorHSL = { h: 0, s: 0, l: 0 };
-  // new THREE.Color(effects.color.toLowerCase()).getHSL(colorHSL);
-  const mat = useMemo(
-    () =>
-      new ShaderMaterial({
-        uniforms: {
-          color: new Uniform(new THREE.Color(color).toArray()),
-          time: new Uniform(0),
-          resolution: new Uniform(new THREE.Vector2(window.innerWidth, window.innerHeight))
-        },
-        vertexShader: vert,
-        fragmentShader: frag,
-        side: DoubleSide,
-      }),
-    [frag, vert]
-  );
-
-  const limiter = useLimiter(30);
-  useFrame(({ clock }) => {
-    if (mat && limiter.isReady(clock)) {
-      mat.uniforms.time.value = clock.getElapsedTime()/2;
-    }
-  });
-
-  return mat;
-};
-
-const vert = `
-    varying vec3 absPosition;
-    varying vec2 vUv;
-    void main() {
-        vUv = uv;
-        absPosition = position;
-        gl_Position = projectionMatrix * modelViewMatrix * vec4( position, 1.0 );
-    }
-`;
-
-const frag = `
-  #define fogNear 0.
-  #define fogFar 1000.
-  #define fogColor vec3(0., 0., 0.)
-
-  uniform highp float time;
-  uniform sampler2D tex;
-  uniform sampler2D tex2;
-  uniform vec2 resolution;
-  uniform vec3 color;
-  
-  varying vec2 vUv;
-  precision mediump float;
-  
-  // vec2 resolution = vec2(2.);
-
-  float random(vec2 v) {
-    return fract(sin(v.x * 32.1231 - v.y * 2.334 + 13399.2312) * 2412.32312);
-  }
-  float random(float x, float y) {
-    return fract(sin(x * 32.1231 - y * 2.334 + 13399.2312) * 2412.32312);
-  }
-  float random(float x) {
-    return fract(sin(x * 32.1231 + 13399.2312) * 2412.32312);
-  }
-  
-  vec3 rgb2hsv(vec3 c) {
-    vec4 K = vec4(0.0, -1.0 / 3.0, 2.0 / 3.0, -1.0);
-    vec4 p = mix(vec4(c.bg, K.wz), vec4(c.gb, K.xy), step(c.b, c.g));
-    vec4 q = mix(vec4(p.xyw, c.r), vec4(c.r, p.yzx), step(p.x, c.r));
-
-    float d = q.x - min(q.w, q.y);
-    float e = 1.0e-10;
-    return vec3(abs(q.z + (q.w - q.y) / (6.0 * d + e)), d / (q.x + e), q.x);
-  }
-
-  float hue2rgb(float f1, float f2, float hue) {
-    if (hue < 0.0)
-        hue += 1.0;
-    else if (hue > 1.0)
-        hue -= 1.0;
-    float res;
-    if ((6.0 * hue) < 1.0)
-        res = f1 + (f2 - f1) * 6.0 * hue;
-    else if ((2.0 * hue) < 1.0)
-        res = f2;
-    else if ((3.0 * hue) < 2.0)
-        res = f1 + (f2 - f1) * ((2.0 / 3.0) - hue) * 6.0;
-    else
-        res = f1;
-    return res;
-  }
-  vec3 hsl2rgb(vec3 hsl) {
-    vec3 rgb;
-    
-    if (hsl.y == 0.0) {
-        rgb = vec3(hsl.z); // Luminance
-    } else {
-        float f2;
-        
-        if (hsl.z < 0.5)
-            f2 = hsl.z * (1.0 + hsl.y);
-        else
-            f2 = hsl.z + hsl.y - hsl.y * hsl.z;
-            
-        float f1 = 2.0 * hsl.z - f2;
-        
-        rgb.r = 0.0; // hue2rgb(f1, f2, hsl.x + (1.0/3.0));
-        rgb.g = cos( hue2rgb(f1, f2, hsl.x ));
-        rgb.b = 0.0; // hue2rgb(f1, f2, hsl.x - (1.0/3.0));
-    }   
-    return rgb;
-  }
-
-  float character(float i) {    
-     return i<15.01? floor(random(i)*32768.) : 0.;
-  }
-
-  void main() {
-    vec2 S = 25. * vec2(3., 50.);
-    // vec2 coord = vec2(
-    //     gl_FragCoord.x / resolution.y,
-    //     gl_FragCoord.y / resolution.y + (resolution.y - resolution.x) / (9. * resolution.y)
-    // );
-    vec2 c = floor(vUv * S);
-
-    float offset = random(c.x) * S.x;
-    float speed = random(c.x * 3.) * .5 + 0.2;
-    float len = random(c.x) * 15. + 100.;
-    float u = 1. - fract(c.y / len + time * speed + offset) * 2.;
-
-    float padding = 2.;
-    vec2 smS = vec2(3., 5.);
-    vec2 sm = floor(fract(vUv * S) * (smS + vec2(padding))) - vec2(padding);
-    float symbol = character(floor(random(c + floor(time * speed)) * 15.));
-    bool s = sm.x < 0. || sm.x > smS.x || sm.y < 0. || sm.y > smS.y ? false
-             : mod(floor(symbol / pow(2., sm.x + sm.y * smS.x)), 2.) == 1.;
-
-    // vec3 curRGB = hsl2rgb(vec3(c.x / S.x, 1., 0.5));
-    vec3 curRGB = color;
-    if( s )
-    {
-        if( u > 0.9 )
-            {
-            curRGB.r = 1.0;
-            curRGB.g = 1.0;
-            curRGB.b = 1.0;
-            }
-        else
-            curRGB = curRGB * u;
-    }
-    else
-        curRGB = vec3( 0.0, 0.0, 0.0 );
-
-    gl_FragColor = vec4(curRGB.x, curRGB.y, curRGB.z, 1.0);
-    
-    // account for fog
-    float depth = gl_FragCoord.z / gl_FragCoord.w;
-    float fogFactor = smoothstep( fogNear, fogFar, depth );
-    gl_FragColor.rgb = mix( gl_FragColor.rgb, fogColor, fogFactor );
-  }
-`;
