+++ src/Realm/properties/RealmScene/Matrix/index.tsx
@@ -1,5 +1,4 @@
 import { useMatrixMat } from "./components/matrix";
-import { useMatrixMat2 } from "./components/matrix2";
 import { useParticleMaterial } from "./components/particleMat";
 import {useLoader} from "@react-three/fiber";
 import * as THREE from "three";
@@ -8,8 +7,7 @@ import {MeshBasicMaterial} from "three";
 
 export default function Matrix() {
 
-  const matrixMat = useMatrixMat(0.456);
-  const matrixMat2 = useMatrixMat2();
+  const matrixMat = useMatrixMat("green");
   const particleMat = useParticleMaterial();
 
   const tex = useLoader(THREE.TextureLoader, "https://d1p3v0j4bqcb21.cloudfront.net/images/matrix.png")
@@ -17,7 +15,7 @@ export default function Matrix() {
   return (
     <group position-z={-2}>
       {/*<Box args={[1,1,1]} material={new MeshBasicMaterial({color: new THREE.Color("blue")})} />*/}
-      <mesh material={matrixMat2}>
+      <mesh material={matrixMat}>
         <planeBufferGeometry args={[10, 10]} />
         {/*<meshBasicMaterial map={tex} />*/}
       </mesh>
